Move inline_functions to dask.optimize
