+++ dask/tests/test_async.py
@@ -80,60 +80,3 @@ def test_state_to_networkx():
     g = state_to_networkx(dsk, state)
     assert isinstance(g, nx.DiGraph)
 
-
-def double(x):
-    return x * 2
-
-def test_inline():
-    x, y, i, d = 'xyid'
-    dsk = {'out': (add, i, d),
-           i: (inc, x),
-           d: (double, y),
-           x: 1, y: 1}
-
-    result = inline_functions(dsk, fast_functions=set([inc]))
-    expected = {'out': (add, (inc, x), d),
-                d: (double, y),
-                x: 1, y: 1}
-    assert result == expected
-
-
-def test_inline_ignores_curries_and_partials():
-    dsk = {'x': 1, 'y': 2,
-           'a': (partial(add, 1), 'x'),
-           'b': (inc, 'a')}
-
-    result = inline_functions(dsk, fast_functions=set([add]))
-    assert 'a' not in set(result.keys())
-
-
-def test_inline_doesnt_shrink_fast_functions_at_top():
-    dsk = {'x': (inc, 'y'), 'y': 1}
-    result = inline_functions(dsk, fast_functions=set([inc]))
-    assert result == dsk
-
-
-def test_inline_traverses_lists():
-    x, y, i, d = 'xyid'
-    dsk = {'out': (sum, [i, d]),
-           i: (inc, x),
-           d: (double, y),
-           x: 1, y: 1}
-    expected = {'out': (sum, [(inc, x), d]),
-                d: (double, y),
-                x: 1, y: 1}
-    result = inline_functions(dsk, fast_functions=set([inc]))
-    assert result == expected
-
-
-def test_functions_of():
-    a = lambda x: x
-    b = lambda x: x
-    c = lambda x: x
-    assert functions_of((a, 1)) == set([a])
-    assert functions_of((a, (b, 1))) == set([a, b])
-    assert functions_of((a, [(b, 1)])) == set([a, b])
-    assert functions_of((a, [[[(b, 1)]]])) == set([a, b])
-    assert functions_of(1) == set()
-    assert functions_of(a) == set()
-    assert functions_of((a,)) == set([a])
